Switch to BigInteger to support int64 IDs
Run this SQL query to alter the types of the columns:
ALTER TABLE afk_users
ALTER COLUMN user_id TYPE BIGINT;
ALTER TABLE antiflood
ALTER COLUMN user_id TYPE BIGINT;
ALTER TABLE connection
ALTER COLUMN user_id TYPE BIGINT;
ALTER TABLE feds_settings
ALTER COLUMN user_id TYPE BIGINT;
ALTER TABLE gbans
ALTER COLUMN user_id TYPE BIGINT;
ALTER TABLE user_report_settings
ALTER COLUMN user_id TYPE BIGINT;
ALTER TABLE userinfo
ALTER COLUMN user_id TYPE BIGINT;
ALTER TABLE users
ALTER COLUMN user_id TYPE BIGINT;
ALTER TABLE warns
ALTER COLUMN user_id TYPE BIGINT;
ALTER TABLE human_checks
ALTER COLUMN user_id TYPE BIGINT;
ALTER TABLE chat_members
ALTER COLUMN "user" TYPE BIGINT;

+++ tg_bot/modules/sql/connection_sql.py
@@ -2,7 +2,7 @@ import threading
 import time
 from typing import Union
 
-from sqlalchemy import Column, String, Boolean, UnicodeText, Integer
+from sqlalchemy import Column, String, Boolean, UnicodeText, Integer, BigInteger
 
 from tg_bot.modules.sql import SESSION, BASE
 
@@ -24,7 +24,7 @@ class ChatAccessConnectionSettings(BASE):
 
 class Connection(BASE):
     __tablename__ = "connection"
-    user_id = Column(Integer, primary_key=True)
+    user_id = Column(BigInteger, primary_key=True)
     chat_id = Column(String(14))
 
     def __init__(self, user_id, chat_id):
@@ -34,7 +34,7 @@ class Connection(BASE):
 
 class ConnectionHistory(BASE):
     __tablename__ = "connection_history"
-    user_id = Column(Integer, primary_key=True)
+    user_id = Column(BigInteger, primary_key=True)
     chat_id = Column(String(14), primary_key=True)
     chat_name = Column(UnicodeText)
     conn_time = Column(Integer)
