+++ tg_bot/modules/sql/feds_sql.py
@@ -1,7 +1,7 @@
 import ast
 import threading
 
-from sqlalchemy import Column, String, UnicodeText, func, distinct, Integer, Boolean
+from sqlalchemy import Column, String, UnicodeText, func, distinct, BigInteger, Boolean
 
 from tg_bot.modules.sql import SESSION, BASE
 
@@ -49,7 +49,7 @@ class BansF(BASE):
 
 class FedsUserSettings(BASE):
 	__tablename__ = "feds_settings"
-	user_id = Column(Integer, primary_key=True)
+	user_id = Column(BigInteger, primary_key=True)
 	should_report = Column(Boolean, default=True)
 
 	def __init__(self, user_id):
